Bump to Magnesium dependencies
https://docs.opendaylight.org/projects/integration-distribution/en/latest/platform-versions.html
- upgrade pom to use ODL Magnesium target dev dependecies

https://wiki.opendaylight.org/view/Magnesium_platform_upgrade
- force jsr305 version
- disable SpotBugs enforcement in artifacts when needed
- UintXX yangtools migration

https://git.opendaylight.org/gerrit/c/mdsal/+/83960
- migrate class BindingStreamEventWriter
      from org.opendaylight.yangtools.yang.binding package
      to   org.opendaylight.mdsal.binding.dom.codec.api

- unforce maven-bundle-plugin version in api/pom.xml
  and ordmodels/device/pom.xml build sections
- unfoce iana-afn-safi version in ordmodels/device/pom.xml dependencies

TODO:
for next changes to come in the release:
- migrate Junit to yangtools UintXX
- propose fixes to reactivate SpotBugs enforcement
  and get rid of "deprecated marked for removal"

JIRA: TRNSPRTPCE-198

+++ pce/src/main/java/org/opendaylight/transportpce/pce/networkanalyzer/PceNode.java
@@ -10,8 +10,8 @@ package org.opendaylight.transportpce.pce.networkanalyzer;
 
 import java.util.List;
 import java.util.Map;
-
 import org.opendaylight.yang.gen.v1.urn.ietf.params.xml.ns.yang.ietf.network.rev180226.NodeId;
+import org.opendaylight.yangtools.yang.common.Uint16;
 
 public interface PceNode {
     String getPceNodeType();
@@ -34,7 +34,7 @@ public interface PceNode {
 
     boolean checkWL(long index);
 
-    Map<String, List<Integer>> getAvailableTribPorts();
+    Map<String, List<Uint16>> getAvailableTribPorts();
 
-    Map<String, List<Integer>> getAvailableTribSlots();
+    Map<String, List<Uint16>> getAvailableTribSlots();
 }
